requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.4.5
PyJWT==2.8.0
python-dotenv
Pillow==10.2.0
requests==2.31.0
